refactor: make a ref counted cell wasm implementation
The previous implementation is not Send + Sync, which in WASM doesn't
matter but makes some libraries impossible to use. This "cheats" by
declaring our Rc<RefCell<>> cell as being Send + Sync in WASM, and have
a non-WASM implementation that uses regular Arc<Mutex<>>.

## BREAKING CHANGE
This is backward compatible, but technically should increment the minor
as the type traits impl changed. The CallFuture<> is now Send+Sync.