fix: allow owned leaves in the certified map hash trees
Currently, all the values in the hash trees produced by the certified
map must be references.  This makes it harder for the clients to
build hash trees on the fly from the contents stored in the map.

This change modifies the definition of the HashTree to allow both
references and values to be stored in the tree.

This is a BREAKING CHANGE, some client code might need to be modified
and recompiled.

Fixes #121.

Actual author: Roman Kashitsyn <roman@dfinity.org>