refactor: move pure assets certification code into a module (#252)
This change is an extensive refactoring of the ic-certified-assets package.
The motivation for this change is to make the code more testable and
organized before adding caching logic.

The general idea behind the refactoring is to split out the "state
machine" logic into pure code that does not depend on Rust CDK and
accepts all the environment (time, certificates, etc.) as function
parameters. This logic moved to state_machine.rs.

lib.rs now contains the minimal glue to connect the state machine to
the Rust CDK interface. Eventually, this part should move back to
https://github.com/dfinity/certified-assets, making
the ic-certified-assets package independent from Rust CDK.

The refactoring does not affect the certification logic; all we do is
moving the code around, improving memory-correctness guarantees (no more
RefCells in the state), and adding tests.