feat: CallFutureRaw - call commit point clarification
Some system-api calls can only be called within a callback. This feature lets the canister know if the code following the call_future.await is in a callback/separate-execution-message or in the same execution-message.
This feature adds the CallFutureRaw type for the return type of the call_raw and call_raw128 functions.
By checking the call_perform_status_code field or the will_perform() method on the CallFuture struct, the canister can know if the code following the .await is within a callback or not.