chore: Add safety comments (#248)
It is good practice, with associated lints, to put `# Safety` on any unsafe function, prefix any unsafe block with `// SAFETY:`, and use explicit unsafe blocks even in unsafe functions, to document where unsafety is, and what true facts about the code are sufficient proof that the operation is safe. This PR adds safety docs to existing unsafe code (and marks some incorrect code to be fixed in a future PR).